lf commit: gate
